[Telemetry] Attempt to fix a flake observed on mac.
This was:
  File "src/tools/perf/../telemetry/telemetry/core/browser.py", line 293, in Close
    self._browser_backend.Close()
  File "src/tools/perf/../telemetry/telemetry/core/backends/chrome/desktop_browser_backend.py", line 223, in Close
    self._proc.terminate()
  File "/Library/Frameworks/Python.framework/Versions/2.7/lib/python2.7/subprocess.py", line 1532, in terminate
    self.send_signal(signal.SIGTERM)
  File "/Library/Frameworks/Python.framework/Versions/2.7/lib/python2.7/subprocess.py", line 1527, in send_signal
    os.kill(self.pid, sig)
OSError: [Errno 3] No such process

It seems like we should avoid trying to terminate the process if it is closed.

BUG=

Review URL: https://codereview.chromium.org/71123002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@235004 0039d316-1c4b-4281-b951-d872f2087c98

+++ telemetry/telemetry/core/backends/chrome/desktop_browser_backend.py
@@ -220,12 +220,13 @@ class DesktopBrowserBackend(chrome_browser_backend.ChromeBrowserBackend):
         return self._proc.poll() != None
 
       # Try to politely shutdown, first.
-      self._proc.terminate()
-      try:
-        util.WaitFor(IsClosed, timeout=1)
-        self._proc = None
-      except util.TimeoutException:
-        pass
+      if not IsClosed():
+        self._proc.terminate()
+        try:
+          util.WaitFor(IsClosed, timeout=1)
+          self._proc = None
+        except util.TimeoutException:
+          pass
 
       # Kill it.
       if not IsClosed():
